Made run_ecomp.sh more verbose regarding errors

+++ bin/run_ecomp.sh
@@ -23,21 +23,43 @@ fi
 
 ECOMP_DIR=`dirname $0`/../
 
-SVN_REVISION=`svn log -r 1:HEAD -l 1 $SVN_REPO | tail -n+2 | head -n 1 | awk '{ print substr($1, 2) }'`
+SVN_REVISION=$(svn log -r 1:HEAD -l 1 $SVN_REPO | tail -n+2 | head -n 1 | awk '{ print substr($1, 2) }')
+if [ -z "$SVN_REVISION" ]; then
+	echo Failed to retrieve SVN start revision from SVN log
+	exit -1
+fi
+
+SVN_INFOOUTPUT=$(svn info $SVN_REPO)
+
+if [ -z "$SVN_INFOOUTPUT" ]; then
+	echo Failed to retrieve SVN end revision from SVN info
+	exit -1
+fi
+
 END_SVN_REVISION=$(svn info $SVN_REPO | grep "^Last Changed Rev: " | awk '{ print $4 }')
 TRANSLATE_AUTHOR=`which translate_bbc_username.pl`
 REPORT_DIRECTORY=`pwd`/$PROJECT_DIR/reports
 
 if [ ! -d "$PROJECT_DIR" ]; then
-  git svn clone --authors-prog $TRANSLATE_AUTHOR -r $SVN_REVISION:$END_SVN_REVISION $SVN_REPO $PROJECT_DIR
-  cd $PROJECT_DIR
+	echo Performing initial git svn clone for revision range $SVN_REVISION - $END_SVN_REVISION
+	git svn clone --authors-prog $TRANSLATE_AUTHOR -r $SVN_REVISION:$END_SVN_REVISION $SVN_REPO $PROJECT_DIR
+	if [ $? != 0 ]; then
+		echo There were errors checking out the project, please fix this before continuing
+		echo The command used to checkout the repository was:
+		echo git svn clone --authors-prog $TRANSLATE_AUTHOR -r $SVN_REVISION:$END_SVN_REVISION $SVN_REPO $PROJECT_DIR
+		exit -1
+	fi
+	cd $PROJECT_DIR
 else
-  cd $PROJECT_DIR
-  git svn rebase --authors-prog $TRANSLATE_AUTHOR
-fi
-
-if [ ! -d "reports" ]; then
-  mkdir "reports"
+	cd $PROJECT_DIR
+	git svn rebase --authors-prog $TRANSLATE_AUTHOR
+	if [ $? != 0 ]; then
+  		echo There were errors updating the project, please fix this before continuing
+  		echo The command used to update the project was:
+  		echo git svn rebase --authors-prog $TRANSLATE_AUTHOR
+  		echo This was ran in the $PROJECT_DIR directory
+		exit -1;
+	fi
 fi
 
 echo -n "Running metrics for $FILE_GLOB files"
@@ -48,4 +70,7 @@ else
 fi
 
 metrics $REPORT_DIRECTORY "$FILE_GLOB" $FORMATTED_EXCLUSIONS
-ln -sf $REPORT_DIRECTORY $ECOMP_DIR/public/data/$PROJECT_DIR
+if [ ! -e "$ECOMP_DIR/public/data/$PROJECT_DIR" ]; then
+	echo Creating project report data symlink from $REPORT_DIRECTORY to $ECOMP_DIR/public/data/$PROJECT_DIR
+	ln -sf $REPORT_DIRECTORY $ECOMP_DIR/public/data/$PROJECT_DIR
+fi
